feat: add profile

+++ server/controllers/userController.js
@@ -91,15 +91,37 @@ class UserController {
 
   async logout(req, res, next) {
     try {
-      console.log('Выход пользователя:', req.body);
-      const { refreshToken } = req.body;
+      const { refreshToken } = req.body; // Получаем токен из тела запроса
+  
+      if (!refreshToken) {
+        return next(ApiError.unauthorized('Токен отсутствует'));
+      }
+  
       await tokenService.removeToken(refreshToken);
-      res.status(200).json({ message: 'Вы успешно вышли из системы' });
+  
+      return res.status(200).json({ message: 'Вы успешно вышли из системы' });
     } catch (error) {
       console.error('Ошибка выхода:', error);
       next(ApiError.internal('Ошибка выхода из системы'));
     }
   }
+  
+  async getUserInfo(req, res, next) {
+    try {
+      const userId = req.params.id; // Получаем ID пользователя из параметров
+      const user = await User.findByPk(userId); // Находим пользователя в базе данных
+  
+      if (!user) {
+        return next(ApiError.notFound('Пользователь не найден'));
+      }
+  
+      res.status(200).json(user); // Возвращаем данные пользователя
+    } catch (error) {
+      console.error('Ошибка получения информации о пользователе:', error);
+      next(ApiError.internal('Ошибка получения информации о пользователе'));
+    }
+  }
+  
 }
 
 module.exports = new UserController();
